+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### Fixed
+ - [#3] (https://github.com/projectapheleia/avl-axi/issues/3) AVL_AXI coverage init issue
+ - [#4] (https://github.com/projectapheleia/avl-axi/issues/4) AXI5 ATOPs should always respond on the B channel
+ - [#5] (https://github.com/projectapheleia/avl-axi/issues/5) Monitors blocked awiting on reset
+
 [v0.2.0] - 2025-10-12
 
 ### Fixed